refactor(core): move ToolEnricher trait to devboy-core, add ProviderMetadata (#66)
- ToolEnricher trait + ToolSchema + sanitize_field_name moved to devboy-core
  (allows provider crates to implement enrichers without circular deps)
- ProviderMetadata added to AdditionalContext (serde_json::Value-based,
  provider enrichers deserialize their own metadata)
- devboy-executor re-exports from core, keeps PipelineFormatEnricher
- Removed transform_output from trait (not needed for current enrichers)
